requirements.txt
black==19.10b0
Django==3.0.4
django-environ==0.4.5
django-extensions==2.2.9
django-redis==4.11.0
djangorestframework==3.11.0
factory-boy==2.12.0
flake8==3.7.9
psycopg2==2.8.4
pytest==5.4.1
pytest-django==3.8.0
pytest-mock==2.0.0
